refactor(Coin.java): Coin.java 위치 변경
Coin.java는 모델이기 떄문에 models 패키지로 위치 변경

BREAKING CHANGE: Coin.java 위치 변경

as-is:
src/main/java/vendingmachine/Coin.java

to-be:
src/main/java/vendingmachine/models/Coin.java
